+++ CHANGELOG.md
@@ -1,6 +1,23 @@
 # Changelog
 All notable changes to this project will be documented in this file.
 
+## [1.14.8] - 2024-02-14
+
+### Added
+
+ - Audit logs stored in database
+ - Column encrypted using AES/GCM/NoPadding
+ - New endpoint for fetching audit events
+ - requires to set environment variable with AES256-SECRET-KEY used in encryption algorithm
+ 
+### Changed
+
+ - Using JFrog for IDS dependency management
+ 
+### Removed
+
+ - Audit events are not stored into file with this change but in DB (H2 as default)
+
 ## [1.14.7] - 2024-01-19
 
 ### Added